Factored out calcAvgRating.
